refactor: move modeling related modules to subpackage
