+++ src/model/modeling/gnomic_helpers.py
