+++ tests/unit/test_model_helpers.py
@@ -14,8 +14,8 @@
 
 import pytest
 
-from model.cobra_helpers import get_unique_metabolite
 from model.exceptions import NoIDMapping
+from model.modeling.cobra_helpers import get_unique_metabolite
 
 
 def test_existing_metabolite(iJO1366):
