+++ tests/unit/test_update_models.py
@@ -17,7 +17,7 @@ import os
 import pytest
 from cobra.io import read_sbml_model
 
-from model.adapter import get_unique_metabolite
+from model.modeling.adapter import get_unique_metabolite
 from tools.update_models import update_local_models
